Construct indices as float64, cast to long.
Seems to be a bug in PyTorch 1.6.0 that's causing the values from the
linspace to look like they're computed as float32 then cast to long.

https://discuss.pytorch.org/t/torch-linspace-precision-issue-in-cuda/97551

Construct them as float64 then cast to long, to get the correct values.

+++ random_scene/autoencoder/dataloader06.py
@@ -250,7 +250,7 @@ class RandomSceneSirenSampleLoader:
             image_size = dims[0] * dims[1]
 
             index_linear = torch.linspace(index_size, index_size+image_size-1, image_size,
-                                          device=sample["sample_count"].device, dtype=torch.long)
+                                          device=sample["sample_count"].device, dtype=torch.float64).to(torch.long)
             index_expanded = torch.repeat_interleave(index_linear, sample["sample_count"], dim=0)
             index_list.append(index_expanded)
             self.dims.append((dims[0], dims[1], index_size))
